+++ CHANGELOG.md
@@ -9,6 +9,9 @@ For help with updating to new Bytewax versions, please see the
 __Add any extra change notes here and we'll put them in the release
 notes on GitHub when we make a new release.__
 
+- {py:obj}`~bytewax.inputs.SimplePollingSource` now allows you to
+  retain state to support at-least-once delivery.
+
 - Fixes a bug when using
   {py:obj}`~bytewax.operators.windowing.SlidingWindower` where values
   would be assigned to an extra window if their timestamps were near